Move view files to flat directory structure
