+++ resources/js/views/Profile.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script setup>
-import { useAuthStore } from '@/stores/auth';
+import { useAuthStore } from '../stores/auth';
 
 const authStore = useAuthStore();
 </script>
