+++ resources/js/views/SettingsIndex.vue
@@ -158,6 +158,6 @@
   </div>
 </template>
 <script setup>
-import { useUserPermissions } from '@/composables/useUserPermission.js';
+import { useUserPermissions } from '../composables/useUserPermission.js';
 const userPermissions = useUserPermissions();
 </script>
