+++ resources/js/views/SettingsRolesView.vue
@@ -272,13 +272,13 @@
 </template>
 
 <script setup>
-import env from '@/env.js';
-import { useApi } from '@/composables/useApi.js';
-import { useFormErrors } from '@/composables/useFormErrors.js';
+import env from '../env.js';
+import { useApi } from '../composables/useApi.js';
+import { useFormErrors } from '../composables/useFormErrors.js';
 import { useRouter } from 'vue-router';
 import { onMounted, ref, computed } from 'vue';
-import { useUserPermissions } from '@/composables/useUserPermission.js';
-import { useSettingsStore } from '@/stores/settings';
+import { useUserPermissions } from '../composables/useUserPermission.js';
+import { useSettingsStore } from '../stores/settings';
 import { useI18n } from 'vue-i18n';
 import _ from 'lodash';
 import ConfirmDialog from 'primevue/confirmdialog';
