+++ resources/js/views/SettingsServersView.vue
@@ -295,16 +295,16 @@
   </div>
 </template>
 <script setup>
-import env from '@/env.js';
-import { useFormErrors } from '@/composables/useFormErrors.js';
-import { useApi } from '@/composables/useApi.js';
-import { useUserPermissions } from '@/composables/useUserPermission.js';
+import env from '../env.js';
+import { useFormErrors } from '../composables/useFormErrors.js';
+import { useApi } from '../composables/useApi.js';
+import { useUserPermissions } from '../composables/useUserPermission.js';
 import { useRouter } from 'vue-router';
 import { useConfirm } from 'primevue/useconfirm';
 import ConfirmDialog from 'primevue/confirmdialog';
 import { useI18n } from 'vue-i18n';
 import { computed, onMounted, ref } from 'vue';
-import { useToast } from '@/composables/useToast.js';
+import { useToast } from '../composables/useToast.js';
 
 const toast = useToast();
 const userPermissions = useUserPermissions();
