+++ resources/js/views/SettingsUsersNew.vue
@@ -176,11 +176,11 @@
   </div>
 </template>
 <script setup>
-import env from '@/env.js';
+import env from '../env.js';
 import 'cropperjs/dist/cropper.css';
-import { useApi } from '@/composables/useApi.js';
-import { useFormErrors } from '@/composables/useFormErrors.js';
-import { useSettingsStore } from '@/stores/settings';
+import { useApi } from '../composables/useApi.js';
+import { useFormErrors } from '../composables/useFormErrors.js';
+import { useSettingsStore } from '../stores/settings';
 import { onMounted, reactive, ref } from 'vue';
 import { useRouter } from 'vue-router';
